renderer: draw the root image as big as it is
Previously we always draw the root image as big as the entire screen
space. Meaning if the root image is small, the outside region will be
clamped to the edge of the root image.

Now, we draw the root image as is, no tiling, no stretching, or clamping
to edges. And since the root image no long covers the entire screen, we
add a clean, black layer underneath so we don't get artifacts when
drawing with damage.

Closes #1464

Changelog: Behavior: The desktop background image is no longer smeared
to fill the screen. If it's too small, it will be too small.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>